+++ VK_SANDBOX/Engine/src/Imgui/ImGuiManager.cpp
@@ -82,7 +82,7 @@ void IHCEngine::IMGUI::ImGuiManager::createImguiDescriptorPool()
     pool_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
     pool_info.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
     pool_info.maxSets = 200;
-    pool_info.poolSizeCount = std::size(pool_sizes);
+    pool_info.poolSizeCount = static_cast<uint32_t>(std::size(pool_sizes));
     pool_info.pPoolSizes = pool_sizes;
 
     auto graphicsManager = IHCEngine::Core::GraphicsManagerLocator::GetGraphicsManager();
